pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trading_bot"
version = "0.1.0"
description = "Advanced trading bot for Bybit"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "ccxt>=4.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "aiohttp>=3.8.5",
    "ta>=0.10.2",
    "scikit-learn>=1.3.0"
]
requires-python = ">=3.8"

[tool.setuptools]
packages = ["trading_bot"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests"
]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
